2025/sch-quals/web/projectdiscovery/app/requirements.txt
Flask
psycopg2-binary
Jinja2
